Install screenshot Gauge plugin on devcontainer
Even though we don't use screenshots (and even though we have now
configured Gauge to not try to take screenshots), Gauge still installs
the screenshot plugin if its not there when running Gauge specs for the
first time.  So we now just install the plugin on the devcontainer to
save that overhead when running the specs.