Only use Gauge spec to add transactions
Prior to this commit we were using Liquibase to seed the database with
three initial transactions. Now that we also have the Gauge spec to
add transactions into the database on demand, it is cleaner to just use
the Gauge spec as the only way in which we add transactions.