Simplify environment variables (#11)
* Simplify required environment variables

Prior to this commit we required a [`LIQUIBASE_COMMAND_URL`][1]
environment variable to be set which required knowing the precise format
(`jdbc:<database>://<host>:<port>/<database_name>;<URL_attributes>`)
that [Liquibase][2] requires.  We have now removed the need for this
environment variable by instead building the Liquibase command URL
programmatically using the simpler `DATABASE_HOST` `DATABASE_PORT` and
`DATABASE_NAME` environment variables (which we were already requiring,
meaning that this commit also removes some duplication).

[1]: https://docs.liquibase.com/workflows/liquibase-community/using-jdbc-url-in-liquibase.html
[2]: https://www.liquibase.org

* Always specify database port via env var

Prior to this commit there was one place in the code where we were
using the default PostgreSQL port (`5432`) rather than externalising it
via our `DATABASE_PORT` environment variable.  (In practice we are
currently always using `5432` as the port, but it's important to be
consistent if this changes, as it could lead to a subtle bug.)

Another reason for this change is that [config should always live
outside the code][1].

[1]: https://12factor.net/config

* Remove duplication of env vars in GitHub Action