Add Dockerfile CI hadolint linting via reviewdog
NB we cannot do Dockerfile linting using Hadolint locally on the
devcontainer as there is an issue installing Hadolint there.  That's ok
though as we can catch any Dockerfile linting errors via the GitHub
Action CI check introduced in this commit, and then fix them on a
feature branch before merging to the `main` branch.

https://github.com/reviewdog/action-hadolint

https://github.com/hadolint/hadolint