Ignore warnings about pinning apt-get versions
Updating pinned versions of apt-get software is cumbersome, as there is
no automatic way to update them (Dependabot doesn't provide a way to do
this, for instance). So we have decided not to pin apt-get versions in
our devcontainer Dockerfile as they are just for the devcontainer and
not for anything production facing.  This commit therefore ignores the
[Hadolint warnings about not pinning apt-get versions][1].

[1]: https://github.com/hadolint/hadolint/wiki/DL3008