Ignore Hadolint about not to use both Wget & curl
We ignore [Hadolint's warnings about not to use both Wget and curl][1],
as they both come preinstalled on our devcontainer.  Amending our bash
commands to just use one of them isn't worth it, especially as the
commands are generally the "standard" way to install the software in
question from the command line.

[1]: https://github.com/hadolint/hadolint/wiki/DL4001