Add Dockerfile CI hadolint linting via reviewdog (#18)
* Add Dockerfile CI hadolint linting via reviewdog

NB we cannot do Dockerfile linting using Hadolint locally on the
devcontainer as there is an issue installing Hadolint there.  That's ok
though as we can catch any Dockerfile linting errors via the GitHub
Action CI check introduced in this commit, and then fix them on a
feature branch before merging to the `main` branch.

https://github.com/reviewdog/action-hadolint

https://github.com/hadolint/hadolint

* Ignore warnings about pinning apt-get versions

Updating pinned versions of apt-get software is cumbersome, as there is
no automatic way to update them (Dependabot doesn't provide a way to do
this, for instance). So we have decided not to pin apt-get versions in
our devcontainer Dockerfile as they are just for the devcontainer and
not for anything production facing.  This commit therefore ignores the
[Hadolint warnings about not pinning apt-get versions][1].

[1]: https://github.com/hadolint/hadolint/wiki/DL3008

* Ignore Hadolint about not to use both Wget & curl

We ignore [Hadolint's warnings about not to use both Wget and curl][1],
as they both come preinstalled on our devcontainer.  Amending our bash
commands to just use one of them isn't worth it, especially as the
commands are generally the "standard" way to install the software in
question from the command line.

[1]: https://github.com/hadolint/hadolint/wiki/DL4001

* Clean up apt-get in Dockerfile as per Hadolint

As per https://github.com/hadolint/hadolint/wiki/DL3009

* Catch any errors in pipe command in Dockerfile

This is as per [Hadolint's linting instruction][1].

[1]: https://github.com/hadolint/hadolint/wiki/DL4006

* Remove brittle cd command in Dockerfile

This is as per [Hadolint's warning][1].

[1]: https://github.com/hadolint/hadolint/wiki/DL3003

* Remove sudo command in Dockerfile

As per [Hadolint's warning][1].

[1]: https://github.com/hadolint/hadolint/wiki/DL3004

* Tidy up Gauge install bash commands in Dockerfile