Fix Gauge spec batching transactions bug
The bank transactions added via the Gauge spec were all being added in
one batch, rather than incrementally every x seconds.  The fix was to
amend the Python indentation, so that each bank transaction is committed
to the database individually.  Lesson learned is that [indentation in
Python is very important][1].

[1]: https://teamtreehouse.com/community/how-does-python-know-where-the-end-of-a-forloop-code-block-is-theres-no-indication-of-when-execution-should-restart