Document AWS storage settings
In particular it's important to document that autoscaling should be
disabled, as with it enabled there's a slight risk that credit card
charges could ensue.

Similarly it's also important to document setting the storage to the
smallest possible size of 5 GiB, so as to minimise the free tier storage
and allow more than one database to be created if need be, for instance.